+++ resource-service/src/main/java/com/microservice/resource/repository/ResourceRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
  * Repository for managing MP3 resources in the database.
  */
 @Repository
-public interface ResourceRepository extends JpaRepository<Resource, Long> {
+public interface ResourceRepository extends JpaRepository<Resource, Integer> {
 }
